pyproject.toml
[tool.poetry]
name = "dosirock"
version = "0.1.0"
description = ""
authors = ["Gomnonix <lee20481997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
python-dotenv = "^1.0.1"

requests = "^2.32.3"
djangorestframework-simplejwt = "^5.3.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
django-stubs = "^5.0.2"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.12
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[[tool.mypy.overrides]]
module = "*.settings.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "dosirock.settings"
